requirements.txt
numpy>=1.18.0
pandas>=1.0.0
matplotlib>=3.1.0
opencv-python>=4.2.0
scipy>=1.4.0
tqdm>=4.45.0
pyyaml>=5.3.0
seaborn>=0.11.0
ultralytics>=8.0.0
